migrate victoria (#9)
Ubuntu support / Master scaling support / Master recreation support (#7)
fix etcd rejoin for masters (#5)
- etcd fix rejoin
- containerd update
- cni update
- delete old drivers
- fix timestamp update problem
Kubernetes 1.27 / Upgrades / Optimizations (#2)
List of changes
Support Kubernetes till 1.27.
Support Fedora CoreOS 38.
Support of Kubernetes upgrade from older versions of magnum.
Support upgrade of Kubernetes cluster services.
Support Autoscaler configuration from nodepools.
Migration to containerd 1.7.2
Deprecation of old resource installation, migrate to "only helm" for cluster services.
Draino deprecation
Dashboard deprecation
keystone-auth deprecation
Etcd upgrade
Auto-scaling upgrade
Auto-healing added
OOCM upgrade
CoreDNS upgrade
Cinder CSI upgrade
Flannel upgrade
Flannel CNI upgrade
Metric server upgrade
Migration to the "registry.k8s.io"
local docker registry
- support kubernetes v1.22
- etcd intervals increased
- cni plugins update 0.9.0> v1.0.0
- flannel-cni plugin update 0.9.0>v1.0
- flannel update v0.13.0>v0.15.0
- coredns update v1.6.6>v1.8.4
- metrics server update v2.8.8>v3.6.0
- allow to get kube scheduler/controller /metrics
- update cinder csi plugin
- keystone auth update to support kuber 1.22
- optimise kubernetes provisioning/upgrade
- local docker/file registry for all services in kubernetes cluster
- fix log_size_max for podman
- autoscaler tag update v1.19.0>v1.21.0
- containerd update (currently not used)
magnum upgrade
- support new versions of kuberentes
- latest flannel
- native cni plugin installiation
- added cinder csi plugin
- added manila csi plugin
- support of fedora cores 33
- drop insecure kubernetes cluster communications
- upgrade of kuberentes clusters
- update prometheus repository

+++ magnum/drivers/common/templates/kubernetes/fragments/enable-cert-api-manager.sh
@@ -1,3 +1,5 @@
+#!/bin/sh
+
 step="enable-cert-api-manager"
 printf "Starting to run ${step}\n"
 
@@ -6,10 +8,11 @@ printf "Starting to run ${step}\n"
 if [ "$(echo "${CERT_MANAGER_API}" | tr '[:upper:]' '[:lower:]')" = "true" ]; then
     cert_dir=/etc/kubernetes/certs
 
-    echo -e "${CA_KEY}" > ${cert_dir}/ca.key
-
-    # chown kube:kube ${cert_dir}/ca.key
-    chmod 400 ${cert_dir}/ca.key
+    # Only write CA key if it doesn't exist
+    if [ ! -f "${cert_dir}/ca.key" ]; then
+        echo -e "${CA_KEY}" > ${cert_dir}/ca.key
+        chmod 400 ${cert_dir}/ca.key
+    fi
 fi
 
 printf "Finished running ${step}\n"
